Append -i option to use time interval between each transaction.
We have added the -i option to specify the millisecond (msec) time interval between transactions.
(The maximum time interval is 10 seconds.)
This is because the transaction simulates a case where the transaction is not executed at high speed in a real situation.

Ex)
When you run 1000 INSERT transactions in Persist journal mode,
To set the time interval between each transaction to 10msec, use the following procedure.

 Mobibench -p / pathname -d 0 -j 2 -i 10

Signed-off-by: sundooKim <sksioi1234@gmail.com>