I modified Makefile to remove SQLite option (SQLITE_DEFAULT_JOURNAL_SIZE_LIMIT)
The existing Mobibench limited the size of the WAL file or Journal file.
The capacity of the old mobile phone was very small.
So, we limited the size of file to 1MB for accurate experiments.
However, since the capacity of the mobile storage has become big enough recently, we removed this option.
Additionally, SQLITE_DEFAULT_JOURNAL_SIZE_LIMIT is to truncate WAL file or Journal file to value I passed as an argument [1].

[1] https://www.sqlite.org/pragma.html

Signed-off-by: sundoo Kim <sksioi12@hanyang.ac.kr>